Including taylor series, removing others
The taylor series reaction included a vector<int>. We already
had a type for that in musica.cpp, but it was bound to the _core module.
The taylor coefficients were in a different pybind11 module and were
only causing an issue on windows. Pybind11 docs say that when you make
an STL type opaque and add bindings to it, this must be done the same for
all modules.