Python refactor, target version 1 of mechanism configuration (#625)
* putting in reaction component, arrhenius, ancillary

* branched

* emission

* first order loss

* photolysis

* ternary chemical

* troe

* user defined

* tunneling, reations

* removing reactions that will no longer be in v1, mechanism parsing slight refactor

* v1 stuff works, still have to correct surface reactions

* phase, phase species, ignoring surface reactions

* updating configurations

* correcting test

* adding required properties for surface reactions

* Including taylor series, removing others

The taylor series reaction included a vector<int>. We already
had a type for that in musica.cpp, but it was bound to the _core module.
The taylor coefficients were in a different pybind11 module and were
only causing an issue on windows. Pybind11 docs say that when you make
an STL type opaque and add bindings to it, this must be done the same for
all modules.

* pulling in corrected test configurations

* syncing some more changes

* addressing copilot pr comments

* addressing PR comment