adding robertson, oregonator tests
These tests show how to check the solve time achieved
Unfortunately, the solve converges and we don't have to call
solve again. We do in micm. The reason it's different is that
the micm tests change the solver parameters to update the
absolute and relative tolerances. We need to make that available to
make these tests require multiple solves, I think.